Create Get Unique Issues
Uses the URLs list (OCR for the "chapter" search) to download metadata for each unique newspaper issue from which those pages came, and constructs a list of URLs to each unique newspaper issue. Getting the entire newspaper issue that might have an installment of fiction instead of just the pages containing "chapter" insures the entire installment in present, avoiding losing chunks of text that run over into the next page.